server/requirements.txt
annotated-types==0.7.0
astroid==3.3.11
beautifulsoup4==4.14.2
blinker==1.9.0
bs4==0.0.2
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.2.1
dill==0.4.0
dnspython==2.7.0
Flask==3.1.1
Flask-Cors==5.0.0
idna==3.10
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mccabe==0.7.0
platformdirs==4.3.8
pydantic==2.11.7
pydantic_core==2.33.2
pylint==3.3.8
pymongo==4.14.0
python-dotenv==1.1.1
requests==2.32.5
soupsieve==2.8
tomlkit==0.13.3
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
pymongo>=4.8.0
python-dotenv>=1.0.0
pydantic>=2.7.0
openai>=1.0.0,<2.0.0
numpy>=1.24.0